+++ tests/component/test_common_api.py
@@ -7,23 +7,73 @@ to a deployment of ska-oso-services in the same cluster
 
 from http import HTTPStatus
 
+import pytest
+
 from . import OSO_SERVICES_URL
 
 
-def test_coordinates_get(authrequests):
+@pytest.mark.parametrize(
+    "identifier, reference_frame, expected_response",
+    [
+        (
+            "47 Tuc",
+            "equatorial",
+            {
+                "equatorial": {
+                    "ra": "00:24:05.3590",
+                    "dec": "-72:04:53.200",
+                    "redshift": 0.0,
+                    "velocity": -17.2,
+                }
+            },
+        ),
+        (
+            "47 Tuc",
+            "galactic",
+            {
+                "galactic": {
+                    "lat": -44.8891,
+                    "lon": 305.9,
+                    "redshift": 0.0,
+                    "velocity": -17.2,
+                }
+            },
+        ),
+        (
+            "M31",
+            "equatorial",
+            {
+                "equatorial": {
+                    "ra": "00:42:44.3300",
+                    "dec": "41:16:07.500",
+                    "redshift": 0.0,
+                    "velocity": -300.0,
+                }
+            },
+        ),
+        (
+            "N10",
+            "galactic",
+            {
+                "galactic": {
+                    "lat": -78.5856,
+                    "lon": 354.21,
+                    "redshift": 0.0,
+                    "velocity": 6800.0,
+                }
+            },
+        ),
+    ],
+)
+def test_coordinates_get(authrequests, identifier, reference_frame, expected_response):
     """
     Test that the GET /coordinates path receives the request
     and returns a success response with the resolved coordinates
     """
 
-    response = authrequests.get(f"{OSO_SERVICES_URL}/coordinates/N10/equatorial")
-    expected_response = {
-        "equatorial": {
-            "dec": "-33:51:30.197",
-            "ra": "00:08:34.5389",
-            "redshift": 0.0,
-            "velocity": 6800.0,
-        }
-    }
+    response = authrequests.get(
+        f"{OSO_SERVICES_URL}/coordinates/{identifier}/{reference_frame}"
+    )
+
     assert response.status_code == HTTPStatus.OK, response.json()
     assert response.json() == expected_response