+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Unreleased
 * [BUGFIX] Set `PGUSER` to `<.Values.global.oda.postgres.database>_admin` if `.Values.global.oda.postgres.user` isn't set
 * Adds SDP configurations to Scheduling Blocks generated from Proposals
 * Changes the SDP TMData source Helm configuration key from `ska-oso-services.rest.sdp.tmdata` to `global.oso.sdp_tmdata`
+* Improvements to target catalog error handling and logging
 
 9.2.0
 ***********